+++ Core/core.h
@@ -5,4 +5,4 @@ float random_float(float min, float max);
 
 int *BubleSortInt(int *array, int array_size, bool direction);
 
-float* BubleSortFloat(float* array, float array_size, bool direction);
+float* BubleSortFloat(float* array, int array_size, bool direction);
